Transfer Traiding Sim to class and added requestSender usage; rename Strategies folder
