+++ HW7/Valeriy_Satkov/Trading/Traiding.ConsoleApp/Strategies/AddShareStrategy.cs
@@ -1,4 +1,4 @@
-﻿namespace Traiding.ConsoleApp.MenuStrategies
+﻿namespace Traiding.ConsoleApp.Strategies
 {
     using System;
     using System.Collections.Generic;
@@ -8,7 +8,7 @@
     using Traiding.ConsoleApp.DependencyInjection;
     using Traiding.ConsoleApp.Dto;
 
-    class EditShareStrategy : IChoiceStrategy
+    class AddShareStrategy : IChoiceStrategy
     {
         public bool CanExecute(string userChoice)
         {
@@ -17,24 +17,14 @@
 
         public string Run(RequestSender requestSender)
         {
-            Console.WriteLine("  Edit Share info service"); // signal about enter into case
+            Console.WriteLine("  Share registration service."); // signal about enter into case
 
-            int id = 0, shareTypeId = 0;
+            int shareTypeId = 0;
             string companyName = string.Empty;
 
             string inputString = string.Empty;
             while (inputString != "e")
             {
-                if (id == 0)
-                {
-                    Console.Write("   Enter the Id of share: ");
-                    inputString = Console.ReadLine();
-                    int inputInt;
-                    int.TryParse(inputString, out inputInt);
-                    if (!StockExchangeValidation.checkId(inputInt)) continue;
-                    id = inputInt;
-                }
-
                 if (string.IsNullOrEmpty(companyName))
                 {
                     Console.Write("   Enter the Company name: ");
@@ -58,24 +48,23 @@
 
             if (inputString == "e")
             {
-                return "Exit from Edit Share info service";
+                return "Exit from registration";
             }
 
             Console.WriteLine("    Wait a few seconds, please.");
 
             var shareInputData = new ShareInputData
             {
-                Id = id,
                 CompanyName = companyName,
                 ShareTypeId = shareTypeId
             };
 
-            var reqResult = requestSender.EditShare(shareInputData);
+            var reqResult = requestSender.AddShare(shareInputData);
             if (string.IsNullOrWhiteSpace(reqResult))
             {
-                return $"Share with Id = {id} was changed! Press Enter.";
+                return "New share was added! Press Enter.";
             }
-            return "Error. Share wasn't changed! Press Enter.";
+            return "Error. Share wasn't added! Press Enter.";
         }
     }
 }
