+++ HW7/Valeriy_Satkov/Trading/Traiding.ConsoleApp/Strategies/DelClientStrategy.cs
@@ -1,4 +1,4 @@
-﻿namespace Traiding.ConsoleApp.MenuStrategies
+﻿namespace Traiding.ConsoleApp.Strategies
 {
     using System;
     using System.Collections.Generic;
@@ -8,16 +8,16 @@
     using Traiding.ConsoleApp.DependencyInjection;
     using Traiding.ConsoleApp.Dto;
 
-    class PrintBalanceZoneByClientIdStrategy : IChoiceStrategy
+    class DelClientStrategy : IChoiceStrategy
     {
         public bool CanExecute(string userChoice)
         {
-            return userChoice.Equals("7");
+            return userChoice.Equals("3");
         }
 
         public string Run(RequestSender requestSender)
         {
-            Console.WriteLine("  Reports service - balance zone"); // signal about enter into case
+            Console.WriteLine("  Remove Clients service."); // signal about enter into case
 
             int clientId = 0;
 
@@ -26,7 +26,7 @@
             {
                 if (clientId == 0)
                 {
-                    Console.Write("   Enter the Id of client: ");
+                    Console.Write("   Enter the Id of client for del: ");
                     inputString = Console.ReadLine();
                     int inputInt;
                     int.TryParse(inputString, out inputInt);
@@ -39,17 +39,17 @@
 
             if (inputString == "e")
             {
-                return "Exit from Reports service - balance zone";
+                return "Exit from Remove Clients service";
             }
 
             Console.WriteLine("    Wait a few seconds, please.");
 
-            var reqResult = requestSender.GetBalanceZoneColor(clientId);
-            if (!string.IsNullOrWhiteSpace(reqResult))
+            var reqResult = requestSender.RemoveClient(clientId);
+            if (string.IsNullOrWhiteSpace(reqResult))
             {
-                return $"Balance zone for client Id = {clientId} — {reqResult}.";
+                return $"     Client with Id = {clientId} was removed! Press Enter.";
             }
-            return "Error. Client wasn't finded! Press Enter.";
+            return "Error. Client wasn't removed! Press Enter.";
         }
     }
 }
