+++ HW7/Valeriy_Satkov/Trading/Traiding.ConsoleApp/Traiding.ConsoleApp.csproj
@@ -85,14 +85,14 @@
     <Compile Include="Dto\ClientInputData.cs" />
     <Compile Include="Logger\ILoggerService.cs" />
     <Compile Include="Logger\LoggerService.cs" />
-    <Compile Include="MenuStrategies\PrintOperationsByClientIdStrategy.cs" />
-    <Compile Include="MenuStrategies\PrintBalanceZoneByClientIdStrategy.cs" />
-    <Compile Include="MenuStrategies\DelShareStrategy.cs" />
-    <Compile Include="MenuStrategies\EditShareStrategy.cs" />
-    <Compile Include="MenuStrategies\AddShareStrategy.cs" />
-    <Compile Include="MenuStrategies\DelClientStrategy.cs" />
-    <Compile Include="MenuStrategies\EditClientStrategy.cs" />
-    <Compile Include="MenuStrategies\AddClientStrategy.cs" />
+    <Compile Include="Strategies\PrintOperationsByClientIdStrategy.cs" />
+    <Compile Include="Strategies\PrintBalanceZoneByClientIdStrategy.cs" />
+    <Compile Include="Strategies\DelShareStrategy.cs" />
+    <Compile Include="Strategies\EditShareStrategy.cs" />
+    <Compile Include="Strategies\AddShareStrategy.cs" />
+    <Compile Include="Strategies\DelClientStrategy.cs" />
+    <Compile Include="Strategies\EditClientStrategy.cs" />
+    <Compile Include="Strategies\AddClientStrategy.cs" />
     <Compile Include="Models\BalanceEntity.cs" />
     <Compile Include="Models\BlockedMoneyEntity.cs" />
     <Compile Include="Models\BlockedSharesNumberEntity.cs" />
@@ -105,7 +105,8 @@
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="StockExchangeValidation.cs" />
-    <Compile Include="MenuStrategies\IChoiceStrategy.cs" />
+    <Compile Include="Strategies\IChoiceStrategy.cs" />
+    <Compile Include="TraidingSimulator.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
