+++ HW7/Valeriy_Satkov/Trading/Traiding.WebAPIConsole/App.config
@@ -7,10 +7,11 @@
   <connectionStrings>
     <!-- name="traidingConnectionString" connectionString="Data Source=localhost;Initial Catalog=StockExchange;" -->
     <!-- name="traidingConnectionString" connectionString="Data Source=localhost;Initial Catalog=StockExchange;Integrated Security=true" -->
+    <!-- name="traidingConnectionString" connectionString="data source=localhost;initial catalog=StockExchange;integrated security=True;MultipleActiveResultSets=True;App=EntityFramework" -->
     <!-- name="traidingConnectionString" connectionString="Data Source=localhost;Initial Catalog=StockExchange;User ID=sa;Password=Qwerty123" -->
     <!-- name="traidingConnectionString" connectionString="Data Source=localhost;Initial Catalog=StockExchange;Persist Security Info=True;User ID=sa;Password=Qwerty123" -->
     <!-- name="traidingConnectionString" connectionString="Data Source=localhost;;AttachDbFilename=|DataDirectory|\StockExchange.mdf;Initial Catalog=StockExchange;Integrated Security=SSPI;" -->
-    <add name="traidingConnectionString" connectionString="Data Source=localhost;Initial Catalog=StockExchange;User ID=sa;Password=Qwerty123"
+    <add name="traidingConnectionString" connectionString="data source=localhost;initial catalog=StockExchange;integrated security=True;MultipleActiveResultSets=True;App=EntityFramework"
          providerName="System.Data.SqlClient" />
   </connectionStrings>
   <startup>
